+++ header.php
@@ -8,7 +8,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.7.0
+ * @version  2.0.0
  */
 
 
@@ -67,6 +67,7 @@ wp_head();
 
 		get_template_part( 'templates/parts/header/site', 'branding' );
 
+		get_template_part( 'templates/parts/menu/menu', 'secondary' );
 		get_template_part( 'templates/parts/menu/menu', 'primary' );
 
 		?>